+++ user-api/src/main/java/zerobase/bud/notification/event/create/CreateCommentEvent.java
@@ -1,13 +1,13 @@
-package zerobase.bud.notification.event;
+package zerobase.bud.notification.event.create;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import zerobase.bud.domain.Member;
-import zerobase.bud.post.domain.QnaAnswer;
+import zerobase.bud.post.domain.Post;
 
 @Getter
 @RequiredArgsConstructor
-public class AddLikeQnaAnswerEvent {
+public class CreateCommentEvent {
     private final Member member;
-    private final QnaAnswer qnaAnswer;
+    private final Post post;
 }
