+++ user-api/src/main/java/zerobase/bud/notification/event/create/CreateQnaAnswerEvent.java
@@ -1,12 +1,13 @@
-package zerobase.bud.notification.event;
+package zerobase.bud.notification.event.create;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import zerobase.bud.domain.Member;
+import zerobase.bud.post.domain.Post;
 
 @Getter
 @RequiredArgsConstructor
-public class FollowEvent {
+public class CreateQnaAnswerEvent {
     private final Member member;
-    private final Member targetMember;
+    private final Post post;
 }
