+++ user-api/src/main/java/zerobase/bud/notification/event/like/AddLikePostEvent.java
@@ -1,0 +1,13 @@
+package zerobase.bud.notification.event.like;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import zerobase.bud.domain.Member;
+import zerobase.bud.post.domain.Post;
+
+@Getter
+@RequiredArgsConstructor
+public class AddLikePostEvent {
+    private final Member member;
+    private final Post post;
+}
