+++ user-api/src/main/java/zerobase/bud/notification/event/like/AddLikeQnaAnswerEvent.java
@@ -1,13 +1,13 @@
-package zerobase.bud.notification.event;
+package zerobase.bud.notification.event.like;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import zerobase.bud.domain.Member;
-import zerobase.bud.post.domain.Post;
+import zerobase.bud.post.domain.QnaAnswer;
 
 @Getter
 @RequiredArgsConstructor
-public class CreateQnaAnswerEvent {
+public class AddLikeQnaAnswerEvent {
     private final Member member;
-    private final Post post;
+    private final QnaAnswer qnaAnswer;
 }
