+++ user-api/src/main/java/zerobase/bud/notification/event/pin/CommentPinEvent.java
@@ -1,16 +1,13 @@
-package zerobase.bud.notification.event;
+package zerobase.bud.notification.event.pin;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import zerobase.bud.comment.domain.Comment;
 import zerobase.bud.domain.Member;
-import zerobase.bud.post.type.PostType;
 
 @Getter
 @RequiredArgsConstructor
-public class AddLikeCommentEvent {
+public class CommentPinEvent {
     private final Member member;
     private final Comment comment;
-    private final PostType postType;
-    private final Long postId;
 }
