+++ user-api/src/main/java/zerobase/bud/notification/event/pin/QnaAnswerCommentPinEvent.java
@@ -1,16 +1,13 @@
-package zerobase.bud.notification.event;
+package zerobase.bud.notification.event.pin;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import zerobase.bud.comment.domain.QnaAnswerComment;
 import zerobase.bud.domain.Member;
-import zerobase.bud.post.type.PostType;
 
 @Getter
 @RequiredArgsConstructor
-public class AddLikeQnaAnswerCommentEvent {
+public class QnaAnswerCommentPinEvent {
     private final Member member;
     private final QnaAnswerComment qnaAnswerComment;
-    private final PostType postType;
-    private final Long postId;
 }
