+++ user-api/src/main/java/zerobase/bud/post/service/PostService.java
@@ -31,8 +31,8 @@ import zerobase.bud.post.dto.PostDto;
 import zerobase.bud.post.dto.SearchMyPagePost;
 import zerobase.bud.post.dto.SearchPost;
 import zerobase.bud.post.dto.UpdatePost;
-import zerobase.bud.notification.event.AddLikePostEvent;
-import zerobase.bud.notification.event.CreatePostEvent;
+import zerobase.bud.notification.event.like.AddLikePostEvent;
+import zerobase.bud.notification.event.create.CreatePostEvent;
 import zerobase.bud.post.repository.*;
 import zerobase.bud.post.type.PostSortType;
 import zerobase.bud.post.type.PostStatus;
