+++ user-api/src/main/java/zerobase/bud/post/service/QnaAnswerService.java
@@ -30,9 +30,9 @@ import org.springframework.web.multipart.MultipartFile;
 import zerobase.bud.awsS3.AwsS3Api;
 import zerobase.bud.common.exception.BudException;
 import zerobase.bud.domain.Member;
-import zerobase.bud.notification.event.AddLikeQnaAnswerEvent;
-import zerobase.bud.notification.event.CreateQnaAnswerEvent;
-import zerobase.bud.notification.event.QnaAnswerPinEvent;
+import zerobase.bud.notification.event.like.AddLikeQnaAnswerEvent;
+import zerobase.bud.notification.event.create.CreateQnaAnswerEvent;
+import zerobase.bud.notification.event.pin.QnaAnswerPinEvent;
 import zerobase.bud.post.domain.Post;
 import zerobase.bud.post.domain.QnaAnswer;
 import zerobase.bud.post.domain.QnaAnswerImage;
