+++ user-api/src/test/java/zerobase/bud/comment/service/CommentServiceTest.java
@@ -31,8 +31,8 @@ import zerobase.bud.comment.repository.CommentRepository;
 import zerobase.bud.common.exception.BudException;
 import zerobase.bud.common.type.ErrorCode;
 import zerobase.bud.domain.Member;
-import zerobase.bud.notification.event.AddLikeCommentEvent;
-import zerobase.bud.notification.event.CommentPinEvent;
+import zerobase.bud.notification.event.like.AddLikeCommentEvent;
+import zerobase.bud.notification.event.pin.CommentPinEvent;
 import zerobase.bud.post.domain.Post;
 import zerobase.bud.post.dto.CommentDto;
 import zerobase.bud.post.repository.PostRepository;
