+++ user-api/src/test/java/zerobase/bud/comment/service/QnaAnswerCommentServiceTest.java
@@ -31,8 +31,8 @@ import zerobase.bud.comment.repository.QnaAnswerCommentRepository;
 import zerobase.bud.common.exception.BudException;
 import zerobase.bud.common.type.ErrorCode;
 import zerobase.bud.domain.Member;
-import zerobase.bud.notification.event.AddLikeQnaAnswerCommentEvent;
-import zerobase.bud.notification.event.QnaAnswerCommentPinEvent;
+import zerobase.bud.notification.event.like.AddLikeQnaAnswerCommentEvent;
+import zerobase.bud.notification.event.pin.QnaAnswerCommentPinEvent;
 import zerobase.bud.post.domain.Post;
 import zerobase.bud.post.domain.QnaAnswer;
 import zerobase.bud.post.dto.QnaAnswerCommentDto;
