+++ user-api/src/test/java/zerobase/bud/post/service/QnaAnswerServiceTest.java
@@ -48,10 +48,10 @@ import zerobase.bud.awsS3.AwsS3Api;
 import zerobase.bud.common.exception.BudException;
 import zerobase.bud.domain.Member;
 import zerobase.bud.notification.domain.NotificationInfo;
-import zerobase.bud.notification.event.AddLikeQnaAnswerEvent;
-import zerobase.bud.notification.event.CreatePostEvent;
-import zerobase.bud.notification.event.CreateQnaAnswerEvent;
-import zerobase.bud.notification.event.QnaAnswerPinEvent;
+import zerobase.bud.notification.event.like.AddLikeQnaAnswerEvent;
+import zerobase.bud.notification.event.create.CreatePostEvent;
+import zerobase.bud.notification.event.create.CreateQnaAnswerEvent;
+import zerobase.bud.notification.event.pin.QnaAnswerPinEvent;
 import zerobase.bud.post.domain.Post;
 import zerobase.bud.post.domain.QnaAnswer;
 import zerobase.bud.post.domain.QnaAnswerImage;
