+++ user-api/src/test/java/zerobase/bud/user/UserServiceTest.java
@@ -12,8 +12,8 @@ import zerobase.bud.common.exception.MemberException;
 import zerobase.bud.common.type.ErrorCode;
 import zerobase.bud.domain.Level;
 import zerobase.bud.domain.Member;
-import zerobase.bud.notification.event.CreatePostEvent;
-import zerobase.bud.notification.event.FollowEvent;
+import zerobase.bud.notification.event.create.CreatePostEvent;
+import zerobase.bud.notification.event.follow.FollowEvent;
 import zerobase.bud.post.repository.PostRepository;
 import zerobase.bud.repository.MemberRepository;
 import zerobase.bud.type.MemberStatus;
